address PR comments and add some typing
